Jestha21/Jestha21/src/main/java/com/demo/repository/UserRepository.java
package com.demo.repository;

import com.demo.entity.User;
import jakarta.persistence.criteria.Predicate;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.stereotype.Repository;

import java.util.Optional;

@Repository
public interface UserRepository extends JpaRepository<User, Long> {
    public Optional<User> findByUsername(String username);

    public User findByEmail(String email);

    // for testing
    @Query("SELECT CASE WHEN COUNT(u) > 0 THEN TRUE ELSE FALSE END FROM User u WHERE u.userID = :id")
    public Boolean isUserExistsById(Integer id);


}
